fix(session): use host & port props for equality check
Remove not set & unused `uri` property and use the `host` and `port`
properties for an equality check instead.

The `getAvailableSessions` function looks for available VSP sessions via
announcement files and checks if the session is already known before
adding it to the list. This check is performed by comparing the `uri`
property of the session instances. However, since this property was always
an empty string, the comparison always returned true, and ViPER only
added one session from the announcement files.